requirements.txt
Django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
python-dotenv>=1.0
psycopg2-binary>=2.9 ; platform_system != 'Windows' # PostgreSQL 사용 시
Pillow>=10.0
asgiref==3.9.1
django-analytical==3.2.0
PyMySQL==1.1.1
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
whitenoise==6.9.0
whitenoise==6.9.0
whitenoise==6.9.0
whitenoise==6.9.0
whitenoise==6.9.0
